☕ Refactored to double to allocate for cents

+++ lib/core/models/category.dart
@@ -3,7 +3,7 @@ import 'expense.dart';
 class ExpenseCategory {
   String name;
   MetaData meta;
-  int budget;
+  double budget;
 
   ExpenseCategory({
     required this.meta,
